codereviewer project - Copy/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary
requests
Werkzeug
markdown2
google-generativeai 